+++ src/app/published-event-schedule/associate/page.tsx
@@ -1,0 +1,13 @@
+import { AssociateEventScheduleForm } from "./associate-event-schedule-form";
+
+function SelectionPublishedEventSchedule() {
+  return (
+    <main>
+      <br />
+      <br />
+      <AssociateEventScheduleForm />
+    </main>
+  );
+}
+
+export default SelectionPublishedEventSchedule;