+++ src/app/published-event-schedule/genesis/page.tsx
@@ -1,0 +1,13 @@
+import { PublishedEventScheduleForm } from "./published-event-schedule-form";
+
+function GenesisPublishedEventSchedule() {
+  return (
+    <main>
+      <br />
+      <br />
+      <PublishedEventScheduleForm />
+    </main>
+  );
+}
+
+export default GenesisPublishedEventSchedule;