+++ src/app/published-event-schedule/page.tsx
@@ -1,0 +1,13 @@
+import { PublishedEventScheduleList } from "./published-event-schedule-list";
+
+function PublishedEventSchedules() {
+  return (
+    <main>
+      <br />
+      <PublishedEventScheduleList />
+      <br />
+    </main>
+  );
+}
+
+export default PublishedEventSchedules;