+++ src/app/published-event-schedule/submit/page.tsx
@@ -1,0 +1,19 @@
+import { ReviewPublishedEventSchedule } from "./review-published-event-schedule";
+import { SubmitPublishedEventScheduleForm } from "./submit-published-event-schedule-form";
+
+function SubmitPublishedEventSchedule() {
+  return (
+    <main>
+      <br />
+      <br />
+      <SubmitPublishedEventScheduleForm />
+      <br />
+      <br />
+      <hr />
+      <br />
+      <ReviewPublishedEventSchedule />
+    </main>
+  );
+}
+
+export default SubmitPublishedEventSchedule;