+++ src/lib/constants.ts
@@ -1,9 +1,13 @@
-export const BASE_URL = "http://localhost:8080";
+export const API_BASE_URL = "http://localhost:8080";
+export const FILE_STORAGE_BASE_URL =
+  "http://localhost:4566/event-admin-service-file-storage";
+
 export const CLIENTS_PATH = "clients";
 export const EVENT_SCHEDULES_PATH = "event-schedules";
 export const EVENTS_PATH = "events";
 export const LOCATIONS_PATH = "locations";
 export const ORGANIZERS_PATH = "organizers";
+export const PUBLISHED_EVENT_SCHEDULES_PATH = "published-event-schedules";
 export const SCHEDULED_EVENTS_PATH = "scheduled-events";
 export const USERS_PATH = "users";