+++ src/lib/features/location/location-api-slice.ts
@@ -1,6 +1,6 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
-import { BASE_URL, CLIENTS_PATH, LOCATIONS_PATH } from "@/lib/constants";
+import { API_BASE_URL, CLIENTS_PATH, LOCATIONS_PATH } from "@/lib/constants";
 import { Location } from "@/lib/features/location/location";
 
 interface LocationEnvelope {
@@ -9,7 +9,7 @@ interface LocationEnvelope {
 }
 
 export const locationApiSlice = createApi({
-  baseQuery: fetchBaseQuery({ baseUrl: `${BASE_URL}/` }),
+  baseQuery: fetchBaseQuery({ baseUrl: `${API_BASE_URL}/` }),
   reducerPath: "locationApi",
   tagTypes: ["Location"],
   endpoints: (build) => ({