+++ src/lib/store.ts
@@ -8,6 +8,8 @@ import { eventScheduleApiSlice } from "@/lib/features/event-schedule/event-sched
 import { eventScheduleSlice } from "@/lib/features/event-schedule/event-schedule-slice";
 import { locationApiSlice } from "@/lib/features/location/location-api-slice";
 import { organizerApiSlice } from "@/lib/features/organizer/organizer-api-slice";
+import { publishedEventScheduleApiSlice } from "@/lib/features/published-event-schedule/published-event-schedule-api-slice";
+import { publishedEventScheduleSlice } from "@/lib/features/published-event-schedule/published-event-schedule-slice";
 import { scheduledEventApiSlice } from "@/lib/features/scheduled-event/scheduled-event-api-slice";
 import { userApiSlice } from "@/lib/features/user/user-api-slice";
 
@@ -21,6 +23,8 @@ const rootReducer = combineSlices(
   eventScheduleSlice,
   locationApiSlice,
   organizerApiSlice,
+  publishedEventScheduleApiSlice,
+  publishedEventScheduleSlice,
   scheduledEventApiSlice,
   userApiSlice
 );
@@ -43,6 +47,7 @@ export const makeStore = () => {
         .concat(eventScheduleApiSlice.middleware)
         .concat(locationApiSlice.middleware)
         .concat(organizerApiSlice.middleware)
+        .concat(publishedEventScheduleApiSlice.middleware)
         .concat(scheduledEventApiSlice.middleware)
         .concat(userApiSlice.middleware);
     },